+++ lib/src/services.dart
@@ -1,4 +1,4 @@
 export 'services/snackbar/snackbar_config.dart';
 export 'services/launcher/launcher.dart';
-export 'services/dialog/dialog_service.dart';
+export 'services/dialog/dialog_service.dart' show DialogPlatform;
 export 'package:url_launcher/url_launcher.dart';